+++ src/app/components/App/App.tsx
@@ -1,5 +1,10 @@
 import { Routes, Route } from "react-router-dom";
-import { SingInPage, SingUpPage, HomePage } from "pages";
+import HomePage from "pages/HomePage";
+import SingInPage from "pages/SignInPage";
+import SingUpPage from "pages/SignUpPage";
+import MyProfilePage from "pages/MyProfilePage";
+import MyProductsPage from "pages/MyProductsPage";
+import { MyAuctionsPage } from "pages/MyAuctionsPage";
 import { Layout, PrivateRoute, PublicRoute } from "app";
 
 import {} from "./App.styled";
@@ -10,7 +15,7 @@ export const App: React.FC = () => {
       <Routes>
         <Route element={<Layout />}>
           <Route
-            path="/sing-in"
+            path="/sign-in"
             element={
               <PublicRoute restricted redirectTo="/">
                 <SingInPage />
@@ -18,7 +23,7 @@ export const App: React.FC = () => {
             }
           />
           <Route
-            path="/sing-up"
+            path="/sign-up"
             element={
               <PublicRoute restricted redirectTo="/">
                 <SingUpPage />
@@ -32,6 +37,30 @@ export const App: React.FC = () => {
               <PublicRoute>
                 <HomePage />
               </PublicRoute>
+            }
+          />
+          <Route
+            path="/my-profile"
+            element={
+              <PrivateRoute redirectTo="/sign-in">
+                <MyProfilePage />
+              </PrivateRoute>
+            }
+          />
+          <Route
+            path="/my-products"
+            element={
+              <PrivateRoute redirectTo="/sign-in">
+                <MyProductsPage />
+              </PrivateRoute>
+            }
+          />
+          <Route
+            path="/my-auctions"
+            element={
+              <PrivateRoute redirectTo="/sign-in">
+                <MyAuctionsPage />
+              </PrivateRoute>
             }
           />
         </Route>
