+++ src/app/components/Layout/Layout.styled.tsx
@@ -1,21 +1,9 @@
 import styled from "@emotion/styled";
-import { CgMenu } from "react-icons/cg";
-
-export const Header = styled.header`
-  width: 100%;
-  height: 120px;
-
-  background-color: #fff;
-`;
 
 export const Container = styled.section`
   width: 1536px;
   margin-right: auto;
   margin-left: auto;
-`;
-
-export const MenuButton = styled(CgMenu)`
-  width: 48px;
 `;
 
 export const Body = styled.section`
