+++ src/app/components/index.ts
@@ -1,3 +1,4 @@
 export * from "./App";
 export * from "./Layout";
 export * from "./Routes";
+export * from "./Header";
